pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vendor-portal"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "asyncpg",
    "itsdangerous",
    "sse-starlette",
    "reportlab",
    "webauthn>=2",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/src"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "pypdf",
]
